requirements.txt
kiteconnect
upstox-python
python-dotenv
schedule
